GH-556 Align variable type plugin with new search type
